register-leo.leox fully working, inc. re-launch elevated
So now both register and unregister work, and we no longer require the
user to manually launch Leo or a CMD shell as administrator. Leo prompts
for elevation at run time.

What's not good: we're using cross-file clones between the scripts
because of the shared request elevation and related code. Have to figure
out how to combine these into a single script, or maybe even move into
Leo itself(?)

Part of #1012

Leo build: 20190314011223