remove links with pyshortcuts now too; [SEE LONG NOTE]
*** WARNING: something weird happened in Leo here ***
look at the node attributes in .leo and .leox. The initials changed from
 `mhw` to `lkj`.

I created remove-desktop-links-pyshort.leox by copying the at-file node
for add-desktop-links-pyshort.leox and renaming the headline. At one
point I had to recopy the contents because Leo __deleted__ the contents
of the add file.

Before that, I created a script-button for add-desktop-links-pyshort.leox.
It seemed to be trying to execute the contents of a different at-file node!
I exited and restarted, and there-after only used Ctrl-B.

Something is screwy in the cache.