New code in sync_tokens now works. Fixed dump problems.
Aha: gen_newline does *not* guarantee newlines. Examples:

if 1: pass
while True: pass
class MyClass: pass
def func: pass
...

As a result, failure to sync on a newline is *not* an error.

Finally, endtoken can be encountered while looking for a missing newline.
In that case, we must allow the endtoken to be rescanned.