Preserve decluttering styles
Leo remembers, which decluttering rules were applied to nodes - it keeps
the data in a dictionary with the key = p.h. The dictionary is kept
as an instance variable 'declutter_patterns' of class LeoQtTree
declared in qt_tree.py.

If a node's header is in that dictionary, Leo gets the decluttered text
and icons from that dictionary, without repeating the check against
the decluttering rules.

Unfortunately styles were not saved in the dictionary. It resulted in
resetting to the default of any decluttering styles on redraw.