Added support for at-indent directive
This directive has an argument (int) which
affects the indentation of the following lines.

This is necessary to allow outlines to represent
efficiently underindented lines in cases where
inserting leo sentinels is not an option. For
example in the python multiline strings, where
inserting sentinels would change the meaning of
the program. This directives are automatically
added during the reading and they're never
written in the derived file.